move from era to process modifier
