+++ PhysicsTools/PatAlgos/python/slimming/slimmedGenJets_cfi.py
@@ -24,5 +24,5 @@ from Configuration.Eras.Modifier_pp_on_AA_2018_cff import pp_on_AA_2018
 from Configuration.Eras.Modifier_pp_on_PbPb_run3_cff import pp_on_PbPb_run3
 (pp_on_AA_2018 | pp_on_PbPb_run3).toModify(slimmedGenJets, src = "ak4HiSignalGenJets")
 (pp_on_AA_2018 | pp_on_PbPb_run3).toModify(slimmedGenJetsAK8, cut = 'pt>9999', nLoose = 0)
-from Configuration.Eras.Modifier_genJetSubEvent_cff import genJetSubEvent
+from Configuration.ProcessModifiers.genJetSubEvent_cff import genJetSubEvent
 genJetSubEvent.toModify(slimmedGenJets, src = "ak4HiGenJetsCleaned")
